requirements.txt
google-generativeai>=0.8.0
numpy>=1.24.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
